+++ web/app/scripts/delta.js
@@ -1,0 +1,10 @@
+// A scalar that represents a change in another scalar.
+'use strict';
+
+angular.module('biggraph').directive('delta', function(util) {
+  return {
+    restrict: 'E',
+    scope: { ref: '=', tip: '=' },
+    templateUrl: 'delta.html',
+  };
+});