+++ app/com/lynxanalytics/biggraph/graph_api/RuntimeContext.scala
@@ -6,6 +6,7 @@ import org.apache.spark
 import com.lynxanalytics.biggraph.graph_util.HadoopFile
 
 case class RuntimeContext(sparkContext: spark.SparkContext,
+                          sqlContext: spark.sql.SQLContext,
                           ioContext: io.IOContext,
                           broadcastDirectory: HadoopFile) {
   // A suitable partitioner for an RDD of N rows.