+++ CHANGELOG.md
@@ -4,7 +4,7 @@ Please add changes to "master", preferably ordered by their significance. (Most 
 
 # Changes
 
- - Numeric fields in CSV files will be imported to LynxKite with the right types.
+ - Numeric fields in CSV files can be imported to LynxKite with the right types.
 
 ### master