+++ CHANGELOG.md
@@ -9,6 +9,7 @@ Please add changes to "master", preferably ordered by their significance. (Most 
  - A collection of goal oriented LynxKite lessons (the "LynxKite Academy") is now available for
    testing. You can find it at `/#/academy`. Any feedback is much appreciated!
  - Fixed data export in Amazon EMR.
+ - Fixed Import JDBC table button.
 
 ### 1.7.2